requirements.txt
azure-core==1.38.2
azure-eventhub==5.15.1
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
